Add admins to config. Massive refactoring
