+++ Makefile
@@ -34,11 +34,11 @@ lint:
 
 .PHONY: build
 build:
-	$(GO) build -v -ldflags '-w $(LDFLAGS)' .
+	$(GO) build -v -ldflags '-w $(LDFLAGS)' ./cmd/grafana-annotations-bot
 
 .PHONY: release
 release:
 	@which gox > /dev/null; if [ $$? -ne 0 ]; then \
 		$(GO) get -u github.com/mitchellh/gox; \
 	fi
-	CGO_ENABLED=0 gox -arch="386 amd64 arm" -verbose -ldflags '-w $(LDFLAGS)' -output="dist/$(EXECUTABLE)-${DRONE_TAG}-{{.OS}}-{{.Arch}}" .
+	CGO_ENABLED=0 gox -arch="386 amd64 arm" -verbose -ldflags '-w $(LDFLAGS)' -output="dist/$(EXECUTABLE)-${DRONE_TAG}-{{.OS}}-{{.Arch}}" ./cmd/grafana-annotations-bot
